refactor(ui): split mesh visualization responsibilities
