+++ crates/mesh-llm-ui/src/features/network/components/MeshViz.tsx
@@ -1,6 +1,5 @@
 import {
   forwardRef,
-  type PointerEvent as ReactPointerEvent,
   useCallback,
   useEffect,
   useImperativeHandle,
@@ -9,16 +8,13 @@ import {
   useRef,
   useState
 } from 'react'
-import { animate, createTimeline, type JSAnimation, type Timeline } from 'animejs'
-import { Maximize2, Minus, Plus, RotateCcw } from 'lucide-react'
+import { animate, type JSAnimation } from 'animejs'
 import { cn } from '@/lib/cn'
 import { isDevelopmentMode } from '@/lib/env'
 import { buildMeshLinks } from '@/features/network/lib/mesh-links'
-import { chooseClusteredMeshNodePosition } from '@/features/network/lib/mesh-placement'
 import {
   MESH_VIZ_DOT_COLOR_SCHEMES,
   meshVizDotColorSchemeAtIndex,
-  nextMeshVizDotColorSchemeIndex,
   themeFromDocument
 } from '@/features/network/lib/mesh-viz-dot-color-schemes'
 import {
@@ -47,21 +43,28 @@ import {
   viewportsMatch
 } from '@/features/network/lib/mesh-viewport'
 import type { MeshNode, Peer, ResolvedTheme } from '@/features/app-tabs/types'
-import { MeshVizDebugControls, type MeshVizGridMode } from '@/features/network/components/MeshVizDebugControls'
+import type { MeshVizGridMode } from '@/features/network/components/MeshVizDebugControls'
 import {
-  createDebugNode,
-  debugNodeMatchesShortcut,
-  debugNodeShortcutCount,
-  getDebugNodeShortcutBlueprint,
-  isTextEditingTarget,
   NODE_LABEL_FADE_THRESHOLD,
   nodeVisuals,
   prefersReducedMotion,
-  type DebugMeshNode,
-  type DebugNodeShortcut
+  type DebugMeshNode
 } from '@/features/network/components/MeshViz.helpers'
-import { MeshVizNode, MeshVizNodeLabel, type MeshVizNodeLifecycle } from '@/features/network/components/MeshVizNode'
+import { type MeshVizNodeLifecycle } from '@/features/network/components/MeshVizNode'
 import { useMeshVizTraffic } from '@/features/network/components/useMeshVizTraffic'
+import { MeshVizCanvas } from '@/features/network/components/MeshVizCanvas'
+import { useMeshVizDebugControls } from '@/features/network/components/useMeshVizDebugControls'
+import {
+  useMeshVizInteractions,
+  type DragState,
+  type PinchZoomState,
+  type TouchPointState
+} from '@/features/network/components/useMeshVizInteractions'
+import {
+  useMeshVizLifecycleAnimations,
+  type LinkRestoreTimelineRecord,
+  type MeshLifecycleTimelineRecord
+} from '@/features/network/components/useMeshVizLifecycleAnimations'
 
 type MeshVizProps = {
   nodes: MeshNode[]
@@ -83,106 +86,12 @@ type MeshVizProps = {
 export type MeshVizHandle = {
   playTraffic: (sourceNodeId: string, targetNodeId: string) => boolean
 }
-type DragState = {
-  active: boolean
-  pointerId: number | null
-  originX: number
-  originY: number
-  panX: number
-  panY: number
-}
-type TouchPointState = {
-  pointerId: number
-  clientX: number
-  clientY: number
-}
-type PinchZoomState = {
-  active: boolean
-  initialDistance: number
-  initialZoom: number
-}
-type MeshLifecycleTimelineRecord = {
-  keys: Set<string>
-  timeline: Timeline
-}
-type LinkRestoreTimelineRecord = {
-  linkIds: Set<string>
-  timeline: Timeline
-}
-
 const VIEWPORT_RECLAMP_DURATION = 220
 const VIEWPORT_RECLAMP_EASE = 'outExpo'
 const RADAR_PULSE_DURATION = 2000
 const RADAR_PULSE_LOOP_DELAY = 1000
 const RADAR_PULSE_EASE = 'linear'
-const NODE_JOIN_STAGGER_MS = 720
-const NODE_JOIN_DURATION_MS = 380
-const LINK_JOIN_DURATION_MS = 420
-const CONNECTED_NODE_PULSE_DELAY_MS = 500
-const CONNECTED_NODE_PULSE_DURATION_MS = 360
-const NODE_JOIN_SETTLE_BUFFER_MS = 60
-const LINK_LEAVE_DURATION_MS = 260
-const NODE_LEAVE_DELAY_MS = 90
-const NODE_LEAVE_DURATION_MS = 280
-const NODE_LEAVE_STAGGER_MS = 260
-const NODE_LEAVE_SETTLE_BUFFER_MS = 120
-const WHEEL_ZOOM_IN = 1.08
-const WHEEL_ZOOM_OUT = 0.92
 const WHEEL_ZOOM_COMMIT_DELAY_MS = 90
-const BUTTON_ZOOM_IN = 1.12
-const BUTTON_ZOOM_OUT = 0.88
-
-function lifecycleTransitionKey(nodeId: string, phase: MeshVizNodeLifecycle) {
-  return `${nodeId}\u001f${phase}`
-}
-
-function isDefined<T>(value: T | undefined): value is T {
-  return value !== undefined
-}
-
-function numericCssValue(value: string | null | undefined, fallback: number) {
-  if (!value || value === 'none') {
-    return fallback
-  }
-
-  const parsedValue = Number.parseFloat(value)
-  return Number.isFinite(parsedValue) ? parsedValue : fallback
-}
-
-function currentElementOpacity(element: Element, fallback: number) {
-  return numericCssValue(
-    element instanceof HTMLElement || element instanceof SVGElement
-      ? element.style.opacity || element.getAttribute('opacity') || window.getComputedStyle(element).opacity
-      : element.getAttribute('opacity'),
-    fallback
-  )
-}
-
-function currentElementScale(element: HTMLElement, fallback: number) {
-  return numericCssValue(element.style.scale || window.getComputedStyle(element).scale, fallback)
-}
-
-function currentStrokeDashOffset(element: SVGElement, fallback: number) {
-  return numericCssValue(
-    element.style.getPropertyValue('stroke-dashoffset') ||
-      element.getAttribute('stroke-dashoffset') ||
-      window.getComputedStyle(element).getPropertyValue('stroke-dashoffset'),
-    fallback
-  )
-}
-
-function nodeJoinSettleDelay(index: number) {
-  return (
-    index * NODE_JOIN_STAGGER_MS +
-    CONNECTED_NODE_PULSE_DELAY_MS +
-    CONNECTED_NODE_PULSE_DURATION_MS +
-    NODE_JOIN_SETTLE_BUFFER_MS
-  )
-}
-
-function nodeLeaveRemovalDelay(index: number) {
-  return index * NODE_LEAVE_STAGGER_MS + NODE_LEAVE_DELAY_MS + NODE_LEAVE_DURATION_MS + NODE_LEAVE_SETTLE_BUFFER_MS
-}
 
 export const MeshViz = forwardRef<MeshVizHandle, MeshVizProps>(function MeshViz(
   {
@@ -717,668 +626,49 @@ export const MeshViz = forwardRef<MeshVizHandle, MeshVizProps>(function MeshViz(
       viewportRef
     })
 
-  useEffect(() => {
-    const previousSnapshots = topologyNodeSnapshotsRef.current
-    const hasTopologySnapshot = hasTopologySnapshotRef.current
-    const currentIds = new Set(currentRenderNodes.map((node) => node.id))
-    const addedNodes = hasTopologySnapshot ? currentRenderNodes.filter((node) => !previousSnapshots.has(node.id)) : []
-    const removedNodes = [...previousSnapshots.entries()]
-      .filter(([nodeId]) => !currentIds.has(nodeId))
-      .map(([, node]) => node)
-
-    if (addedNodes.length > 0 || removedNodes.length > 0) {
-      setNodeLifecyclePhases((current) => {
-        const next = { ...current }
-
-        for (const node of addedNodes) {
-          next[node.id] = reduceMotion ? 'present' : 'entering'
-        }
-
-        for (const node of removedNodes) {
-          if (reduceMotion) {
-            delete next[node.id]
-          } else {
-            next[node.id] = 'leaving'
-          }
-        }
-
-        return next
-      })
-    }
-
-    if (reduceMotion) {
-      for (const node of [...addedNodes, ...removedNodes]) {
-        const existingTimeout = nodeLifecycleTimeoutsRef.current.get(node.id)
-
-        if (existingTimeout !== undefined) {
-          window.clearTimeout(existingTimeout)
-          nodeLifecycleTimeoutsRef.current.delete(node.id)
-        }
-      }
-
-      if (removedNodes.length > 0) {
-        const removedNodeIds = new Set(removedNodes.map((node) => node.id))
-        setExitingNodes((current) => current.filter((node) => !removedNodeIds.has(node.id)))
-      }
-
-      topologyNodeSnapshotsRef.current = new Map(currentRenderNodes.map((node) => [node.id, node]))
-      hasTopologySnapshotRef.current = true
-      return
-    }
-
-    if (addedNodes.length > 0) {
-      setExitingNodes((current) => current.filter((node) => !currentIds.has(node.id)))
-
-      addedNodes.forEach((node, index) => {
-        const existingTimeout = nodeLifecycleTimeoutsRef.current.get(node.id)
-
-        if (existingTimeout !== undefined) {
-          window.clearTimeout(existingTimeout)
-        }
-
-        const timeout = window.setTimeout(() => {
-          nodeLifecycleTimeoutsRef.current.delete(node.id)
-          setNodeLifecyclePhases((current) => {
-            if (current[node.id] !== 'entering') return current
-            return { ...current, [node.id]: 'present' }
-          })
-        }, nodeJoinSettleDelay(index))
-
-        nodeLifecycleTimeoutsRef.current.set(node.id, timeout)
-      })
-    }
-
-    if (removedNodes.length > 0) {
-      setExitingNodes((current) => {
-        const activeExitingNodes = current.filter((node) => !currentIds.has(node.id))
-        const activeExitingIds = new Set(activeExitingNodes.map((node) => node.id))
-        const nextRemovedNodes = removedNodes.filter((node) => !activeExitingIds.has(node.id))
-
-        return [...activeExitingNodes, ...nextRemovedNodes]
-      })
-
-      removedNodes.forEach((node, index) => {
-        const existingTimeout = nodeLifecycleTimeoutsRef.current.get(node.id)
-
-        if (existingTimeout !== undefined) {
-          window.clearTimeout(existingTimeout)
-        }
-
-        const timeout = window.setTimeout(() => {
-          nodeLifecycleTimeoutsRef.current.delete(node.id)
-          setExitingNodes((current) => current.filter((exitingNode) => exitingNode.id !== node.id))
-          setNodeLifecyclePhases((current) => {
-            if (!(node.id in current)) return current
-
-            const { [node.id]: removedPhase, ...next } = current
-            void removedPhase
-            return next
-          })
-        }, nodeLeaveRemovalDelay(index))
-
-        nodeLifecycleTimeoutsRef.current.set(node.id, timeout)
-      })
-    }
-
-    topologyNodeSnapshotsRef.current = new Map(currentRenderNodes.map((node) => [node.id, node]))
-    hasTopologySnapshotRef.current = true
-  }, [currentRenderNodes, reduceMotion])
-
-  useLayoutEffect(() => {
-    const activeTransitionKeys = new Set<string>()
-    const transitioningNodes = renderNodes
-      .map((node) => ({ node, phase: nodeLifecyclePhase(node.id) }))
-      .filter(({ phase }) => phase === 'entering' || phase === 'leaving')
-
-    for (const { node, phase } of transitioningNodes) {
-      activeTransitionKeys.add(lifecycleTransitionKey(node.id, phase))
-    }
-
-    for (const key of [...nodeLifecycleAnimationKeysRef.current]) {
-      if (!activeTransitionKeys.has(key)) {
-        nodeLifecycleAnimationKeysRef.current.delete(key)
-      }
-    }
-
-    for (const [recordId, record] of meshLifecycleTimelineRecordsRef.current) {
-      const stillActive = [...record.keys].some((key) => activeTransitionKeys.has(key))
-
-      if (!stillActive) {
-        record.timeline.pause()
-        meshLifecycleTimelineRecordsRef.current.delete(recordId)
-      }
-    }
-
-    if (reduceMotion) {
-      for (const record of meshLifecycleTimelineRecordsRef.current.values()) {
-        record.timeline.revert()
-      }
-
-      meshLifecycleTimelineRecordsRef.current.clear()
-      nodeLifecycleAnimationKeysRef.current.clear()
-      return undefined
-    }
-
-    if (transitioningNodes.length === 0) {
-      return undefined
-    }
-
-    const nodeLayerElement = nodeLayerRef.current
-    const svgPanLayerElement = svgPanLayerRef.current
-
-    if (!nodeLayerElement || !svgPanLayerElement) {
-      return undefined
-    }
-
-    const nodeCoreElements = new Map<string, HTMLElement>()
-    nodeLayerElement.querySelectorAll<HTMLElement>('[data-mesh-node-core]').forEach((element) => {
-      const nodeId = element.dataset.meshNodeCore
-
-      if (nodeId) {
-        nodeCoreElements.set(nodeId, element)
-      }
-    })
-
-    const linkElements = new Map<string, SVGLineElement>()
-    svgPanLayerElement.querySelectorAll<SVGLineElement>('[data-mesh-link-id]').forEach((element) => {
-      const linkId = element.dataset.meshLinkId
-
-      if (linkId) {
-        linkElements.set(linkId, element)
-      }
-    })
-
-    const newTransitions = transitioningNodes.filter(({ node, phase }) => {
-      const key = lifecycleTransitionKey(node.id, phase)
-
-      return !nodeLifecycleAnimationKeysRef.current.has(key)
-    })
-
-    if (newTransitions.length === 0) {
-      return undefined
-    }
-
-    const timelineId = meshLifecycleTimelineIdRef.current
-    meshLifecycleTimelineIdRef.current += 1
-    const timelineKeys = new Set(newTransitions.map(({ node, phase }) => lifecycleTransitionKey(node.id, phase)))
-    const transitionIndexByNodeId = new Map(newTransitions.map(({ node }, index) => [node.id, index]))
-    const animatedLinkIds = new Set<string>()
-    const enteringCoreRestingShadows = new Map<HTMLElement, string>()
-    const pulsedCoreElements = new Set<HTMLElement>()
-    const enteringLinkElements = new Set<SVGLineElement>()
-
-    for (const key of timelineKeys) {
-      nodeLifecycleAnimationKeysRef.current.add(key)
-    }
-
-    const timeline = createTimeline({
-      defaults: { ease: 'outQuart' },
-      onComplete: () => {
-        for (const [element, boxShadow] of enteringCoreRestingShadows) {
-          element.style.removeProperty('opacity')
-          element.style.removeProperty('scale')
-          element.style.boxShadow = boxShadow
-        }
-
-        for (const element of pulsedCoreElements) {
-          if (!enteringCoreRestingShadows.has(element)) {
-            element.style.removeProperty('opacity')
-            element.style.removeProperty('scale')
-          }
-        }
-
-        for (const element of enteringLinkElements) {
-          element.style.removeProperty('opacity')
-          element.style.removeProperty('stroke-dashoffset')
-        }
-
-        meshLifecycleTimelineRecordsRef.current.delete(timelineId)
-      }
-    })
-
-    newTransitions.forEach(({ node, phase }, index) => {
-      const nodeCoreElement = nodeCoreElements.get(node.id)
-
-      if (!nodeCoreElement) {
-        return
-      }
-
-      const connectedLinks = links.filter((link) => link.source.id === node.id || link.target.id === node.id)
-      const connectedLinkElements = connectedLinks
-        .filter((link) => {
-          if (animatedLinkIds.has(link.id)) {
-            return false
-          }
-
-          const otherNodeId = link.source.id === node.id ? link.target.id : link.source.id
-          const otherTransitionIndex = transitionIndexByNodeId.get(otherNodeId)
-
-          if (otherTransitionIndex === undefined) {
-            return true
-          }
-
-          const otherPhase = nodeLifecyclePhase(otherNodeId)
-
-          if (otherPhase !== phase) {
-            return phase === 'leaving'
-          }
-
-          return phase === 'entering' ? index >= otherTransitionIndex : index <= otherTransitionIndex
-        })
-        .map((link) => {
-          const element = linkElements.get(link.id)
-
-          if (element) {
-            animatedLinkIds.add(link.id)
-          }
-
-          return element
-        })
-        .filter(isDefined)
-
-      if (phase === 'entering') {
-        const start = index * NODE_JOIN_STAGGER_MS
-        const nodeColor = nodeCoreElement.style.color || 'currentColor'
-
-        enteringCoreRestingShadows.set(nodeCoreElement, nodeCoreElement.style.boxShadow)
-
-        timeline.set(
-          nodeCoreElement,
-          {
-            opacity: 0,
-            scale: 0.54,
-            boxShadow: `0 0 0 0 color-mix(in oklab, ${nodeColor} 0%, transparent)`
-          },
-          start
-        )
-        timeline.add(
-          nodeCoreElement,
-          {
-            opacity: [0, 1, 0.98],
-            scale: [0.6, 1.34, 1.08],
-            boxShadow: [
-              `0 0 0 0 color-mix(in oklab, ${nodeColor} 0%, transparent)`,
-              `0 0 30px 3px color-mix(in oklab, ${nodeColor} 34%, transparent)`,
-              `0 0 14px 1px color-mix(in oklab, ${nodeColor} 20%, transparent)`
-            ],
-            duration: NODE_JOIN_DURATION_MS
-          },
-          start
-        )
-
-        if (connectedLinkElements.length > 0) {
-          for (const linkElement of connectedLinkElements) {
-            enteringLinkElements.add(linkElement)
-          }
-
-          timeline.set(connectedLinkElements, { opacity: 0, strokeDashoffset: 1 }, start)
-          timeline.add(
-            connectedLinkElements,
-            {
-              opacity: [0, 0.62],
-              strokeDashoffset: [1, 0],
-              duration: LINK_JOIN_DURATION_MS
-            },
-            start + NODE_JOIN_DURATION_MS
-          )
-        }
-
-        const connectedCoreElements = connectedLinks
-          .map((link) => (link.source.id === node.id ? link.target.id : link.source.id))
-          .filter((nodeId) => nodeLifecyclePhase(nodeId) === 'present')
-          .map((nodeId) => nodeCoreElements.get(nodeId))
-          .filter(isDefined)
-
-        if (connectedCoreElements.length > 0) {
-          for (const element of connectedCoreElements) {
-            pulsedCoreElements.add(element)
-          }
-
-          timeline.add(
-            connectedCoreElements,
-            {
-              opacity: [0.98, 1, 0.98],
-              scale: [1, 1.12, 1],
-              duration: CONNECTED_NODE_PULSE_DURATION_MS
-            },
-            start + CONNECTED_NODE_PULSE_DELAY_MS
-          )
-        }
-
-        return
-      }
-
-      const start = index * NODE_LEAVE_STAGGER_MS
-      const nodeColor = nodeCoreElement.style.color || 'currentColor'
-
-      if (connectedLinkElements.length > 0) {
-        for (const linkElement of connectedLinkElements) {
-          timeline.add(
-            linkElement,
-            {
-              opacity: [currentElementOpacity(linkElement, 0.62), 0],
-              strokeDashoffset: [currentStrokeDashOffset(linkElement, 0), -1],
-              duration: LINK_LEAVE_DURATION_MS,
-              ease: 'inQuart'
-            },
-            start
-          )
-        }
-      }
-
-      timeline.add(
-        nodeCoreElement,
-        {
-          opacity: [currentElementOpacity(nodeCoreElement, 0.98), 0],
-          scale: [currentElementScale(nodeCoreElement, 1.08), 0.72],
-          boxShadow: [
-            nodeCoreElement.style.boxShadow || `0 0 14px 1px color-mix(in oklab, ${nodeColor} 20%, transparent)`,
-            `0 0 0 0 color-mix(in oklab, ${nodeColor} 0%, transparent)`
-          ],
-          duration: NODE_LEAVE_DURATION_MS,
-          ease: 'inQuart'
-        },
-        start + NODE_LEAVE_DELAY_MS
-      )
-    })
-
-    meshLifecycleTimelineRecordsRef.current.set(timelineId, { keys: timelineKeys, timeline })
-
-    return undefined
-  }, [links, nodeLifecyclePhase, reduceMotion, renderNodes])
-
-  useLayoutEffect(() => {
-    const currentLinkIds = new Set(links.map((link) => link.id))
-    const currentLinksById = new Map(links.map((link) => [link.id, link]))
-    const previousLinkIds = previousLinkIdsRef.current
-
-    previousLinkIdsRef.current = currentLinkIds
-
-    for (const linkId of [...linkRestoreAnimationIdsRef.current]) {
-      if (!currentLinkIds.has(linkId)) {
-        linkRestoreAnimationIdsRef.current.delete(linkId)
-      }
-    }
-
-    for (const [recordId, record] of linkRestoreTimelineRecordsRef.current) {
-      const stillRestoring = [...record.linkIds].every((linkId) => {
-        const link = currentLinksById.get(linkId)
-
-        return link && linkLifecyclePhase(link.source.id, link.target.id) === 'present'
-      })
-
-      if (!stillRestoring) {
-        record.timeline.pause()
-
-        for (const linkId of record.linkIds) {
-          linkRestoreAnimationIdsRef.current.delete(linkId)
-        }
-
-        linkRestoreTimelineRecordsRef.current.delete(recordId)
-      }
-    }
-
-    if (reduceMotion) {
-      for (const record of linkRestoreTimelineRecordsRef.current.values()) {
-        record.timeline.revert()
-      }
-
-      linkRestoreTimelineRecordsRef.current.clear()
-      linkRestoreAnimationIdsRef.current.clear()
-      return undefined
-    }
-
-    if (previousLinkIds.size === 0) {
-      return undefined
-    }
-
-    const restoredLinks = links.filter(
-      (link) =>
-        !previousLinkIds.has(link.id) &&
-        !linkRestoreAnimationIdsRef.current.has(link.id) &&
-        linkLifecyclePhase(link.source.id, link.target.id) === 'present'
-    )
-
-    if (restoredLinks.length === 0) {
-      return undefined
-    }
-
-    const svgPanLayerElement = svgPanLayerRef.current
-
-    if (!svgPanLayerElement) {
-      return undefined
-    }
-
-    const linkElements = new Map<string, SVGLineElement>()
-    svgPanLayerElement.querySelectorAll<SVGLineElement>('[data-mesh-link-id]').forEach((element) => {
-      const linkId = element.dataset.meshLinkId
-
-      if (linkId) {
-        linkElements.set(linkId, element)
-      }
-    })
-    const restoredLinkElements = restoredLinks.map((link) => linkElements.get(link.id)).filter(isDefined)
-
-    if (restoredLinkElements.length === 0) {
-      return undefined
-    }
-
-    const timelineId = linkRestoreTimelineIdRef.current
-    linkRestoreTimelineIdRef.current += 1
-
-    const restoredLinkIds = new Set(restoredLinks.map((link) => link.id))
-
-    for (const linkId of restoredLinkIds) {
-      linkRestoreAnimationIdsRef.current.add(linkId)
-    }
-
-    const timeline = createTimeline({
-      defaults: { ease: 'outQuart' },
-      onComplete: () => {
-        for (const element of restoredLinkElements) {
-          element.style.removeProperty('opacity')
-          element.style.removeProperty('stroke-dashoffset')
-        }
-
-        for (const linkId of restoredLinkIds) {
-          linkRestoreAnimationIdsRef.current.delete(linkId)
-        }
-
-        linkRestoreTimelineRecordsRef.current.delete(timelineId)
-      }
-    })
-
-    timeline.set(restoredLinkElements, { opacity: 0, strokeDashoffset: 1 }, 0)
-    timeline.add(
-      restoredLinkElements,
-      {
-        opacity: [0, 0.62],
-        strokeDashoffset: [1, 0],
-        duration: LINK_JOIN_DURATION_MS
-      },
-      0
-    )
-
-    linkRestoreTimelineRecordsRef.current.set(timelineId, { linkIds: restoredLinkIds, timeline })
-
-    return undefined
-  }, [links, linkLifecyclePhase, reduceMotion])
-
-  useEffect(
-    () => () => {
-      for (const timeout of nodeLifecycleTimeoutsRef.current.values()) {
-        window.clearTimeout(timeout)
-      }
-
-      nodeLifecycleTimeoutsRef.current.clear()
-      nodeLifecycleAnimationKeysRef.current.clear()
-      previousLinkIdsRef.current.clear()
-      linkRestoreAnimationIdsRef.current.clear()
-
-      for (const record of meshLifecycleTimelineRecordsRef.current.values()) {
-        record.timeline.revert()
-      }
-
-      meshLifecycleTimelineRecordsRef.current.clear()
-
-      for (const record of linkRestoreTimelineRecordsRef.current.values()) {
-        record.timeline.revert()
-      }
-
-      linkRestoreTimelineRecordsRef.current.clear()
-    },
-    []
-  )
-
-  useEffect(() => {
-    if (openNodeId && !currentRenderNodeIds.has(openNodeId)) {
-      setOpenNodeId(undefined)
-    }
-
-    if (localHoveredNodeId && !currentRenderNodeIds.has(localHoveredNodeId)) {
-      setLocalHoveredNodeId(undefined)
-    }
-  }, [currentRenderNodeIds, localHoveredNodeId, openNodeId])
-
-  const addDebugNode = useCallback(
-    (shortcut: DebugNodeShortcut) => {
-      const blueprint = getDebugNodeShortcutBlueprint(shortcut)
-
-      setDebugNodes((current) => {
-        const debugIndex = debugNodeCounterRef.current + 1
-        const placementNodes: MeshNode[] = [...nodes, ...current]
-        const position = chooseClusteredMeshNodePosition(meshSeed, debugIndex, blueprint, placementNodes)
-        const debugNode = createDebugNode(debugIndex, blueprint, position)
-
-        debugNodeCounterRef.current = debugIndex
-        return [...current, debugNode]
-      })
-    },
-    [meshSeed, nodes]
-  )
-
-  const removeDebugNode = useCallback((shortcut: DebugNodeShortcut) => {
-    setDebugNodes((current) => {
-      let removeIndex = -1
-
-      for (let index = current.length - 1; index >= 0; index -= 1) {
-        if (debugNodeMatchesShortcut(current[index], shortcut)) {
-          removeIndex = index
-          break
-        }
-      }
-
-      if (removeIndex === -1) {
-        return current
-      }
-
-      return current.filter((_, index) => index !== removeIndex)
-    })
-  }, [])
-
-  const cycleDotColorScheme = useCallback(() => {
-    setDotColorSchemeIndex(nextMeshVizDotColorSchemeIndex)
-  }, [])
-
-  const selectDotColorScheme = useCallback((index: number) => {
-    setDotColorSchemeIndex(index)
-  }, [])
+  useMeshVizLifecycleAnimations({
+    currentRenderNodes,
+    currentRenderNodeIds,
+    renderNodes,
+    links,
+    reduceMotion,
+    nodeLifecyclePhase,
+    linkLifecyclePhase,
+    nodeLayerRef,
+    svgPanLayerRef,
+    topologyNodeSnapshotsRef,
+    hasTopologySnapshotRef,
+    nodeLifecycleTimeoutsRef,
+    nodeLifecycleAnimationKeysRef,
+    meshLifecycleTimelineRecordsRef,
+    meshLifecycleTimelineIdRef,
+    previousLinkIdsRef,
+    linkRestoreAnimationIdsRef,
+    linkRestoreTimelineRecordsRef,
+    linkRestoreTimelineIdRef,
+    setNodeLifecyclePhases,
+    setExitingNodes,
+    openNodeId,
+    localHoveredNodeId,
+    setOpenNodeId,
+    setLocalHoveredNodeId
+  })
+
+  const { addDebugNode, cycleDotColorScheme, removeDebugNode, selectDotColorScheme } = useMeshVizDebugControls({
+    debugNodeCounterRef,
+    debugShortcutsEnabled,
+    meshSeed,
+    nodes,
+    playRandomTraffic,
+    playSelfTraffic,
+    setDebugNodes,
+    setDotColorSchemeIndex,
+    setDotColorSchemeTheme,
+    setGridMode,
+    setShowPanBounds
+  })
 
   useImperativeHandle(ref, () => ({ playTraffic }), [playTraffic])
-
-  useEffect(() => {
-    if (typeof document === 'undefined' || typeof MutationObserver === 'undefined') {
-      return undefined
-    }
-
-    const root = document.documentElement
-    const syncTheme = () => setDotColorSchemeTheme(themeFromDocument())
-
-    syncTheme()
-
-    const observer = new MutationObserver(syncTheme)
-    observer.observe(root, { attributes: true, attributeFilter: ['data-theme'] })
-
-    return () => observer.disconnect()
-  }, [])
-
-  useEffect(() => {
-    if (!debugShortcutsEnabled) {
-      return undefined
-    }
-
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.defaultPrevented || isTextEditingTarget(event.target)) {
-        return
-      }
-
-      const key = event.key.toLowerCase()
-      const debugNodeShortcut = debugNodeShortcutCount(event)
-
-      if (!event.ctrlKey && !event.metaKey && !event.altKey && !event.shiftKey) {
-        if (key === 'z') {
-          event.preventDefault()
-          playRandomTraffic()
-          return
-        }
-
-        if (key === 'x') {
-          event.preventDefault()
-          playSelfTraffic()
-          return
-        }
-      }
-
-      if (debugNodeShortcut !== undefined) {
-        if (event.shiftKey && !event.ctrlKey && !event.metaKey && !event.altKey) {
-          event.preventDefault()
-          removeDebugNode(debugNodeShortcut)
-          return
-        }
-
-        if (!event.ctrlKey || event.metaKey || event.altKey || event.shiftKey) {
-          return
-        }
-
-        event.preventDefault()
-        addDebugNode(debugNodeShortcut)
-        return
-      }
-
-      if (key === 'b') {
-        if (!event.ctrlKey || event.metaKey || event.altKey || event.shiftKey) {
-          return
-        }
-
-        event.preventDefault()
-        setShowPanBounds((current) => !current)
-        return
-      }
-
-      if (key === 'g') {
-        if (!event.ctrlKey || event.metaKey || event.altKey || event.shiftKey) {
-          return
-        }
-
-        event.preventDefault()
-        setGridMode((current) => (current === 'line' ? 'dot' : 'line'))
-        return
-      }
-
-      if (key === 'c') {
-        if (!event.ctrlKey || event.metaKey || event.altKey || event.shiftKey) {
-          return
-        }
-
-        event.preventDefault()
-        cycleDotColorScheme()
-      }
-    }
-
-    window.addEventListener('keydown', onKeyDown)
-
-    return () => window.removeEventListener('keydown', onKeyDown)
-  }, [addDebugNode, cycleDotColorScheme, debugShortcutsEnabled, playRandomTraffic, playSelfTraffic, removeDebugNode])
 
   useEffect(() => {
     const canvasElement = canvasRef.current
@@ -1567,234 +857,31 @@ export const MeshViz = forwardRef<MeshVizHandle, MeshVizProps>(function MeshViz(
     }
   }, [fitNodes, updateCanvasSize])
 
-  const touchPoints = () => Array.from(touchPointersRef.current.values())
-
-  const distanceBetweenTouchPoints = (first: TouchPointState, second: TouchPointState) =>
-    Math.hypot(second.clientX - first.clientX, second.clientY - first.clientY)
-
-  const midpointRelativeToCanvas = (element: HTMLDivElement, first: TouchPointState, second: TouchPointState) => {
-    const rect = element.getBoundingClientRect()
-
-    return {
-      x: (first.clientX + second.clientX) * 0.5 - rect.left,
-      y: (first.clientY + second.clientY) * 0.5 - rect.top
-    }
-  }
-
-  const beginPinchZoom = () => {
-    const [first, second] = touchPoints()
-
-    if (!first || !second) {
-      return
-    }
-
-    dragRef.current.active = false
-    dragRef.current.pointerId = null
-    setIsPanning(false)
-    pinchZoomRef.current = {
-      active: true,
-      initialDistance: Math.max(1, distanceBetweenTouchPoints(first, second)),
-      initialZoom: viewportRef.current.zoom
-    }
-  }
-
-  const handleCanvasPointerDown = (event: ReactPointerEvent<HTMLDivElement>) => {
-    if (event.button !== 0) {
-      return
-    }
-
-    if (event.target instanceof Element && event.target.closest('button, a, input, label')) {
-      return
-    }
-
-    event.preventDefault()
-    viewportAnimationRef.current?.revert()
-    viewportAnimationRef.current = undefined
-    hasUserControlledViewportRef.current = true
-    pendingPanTransformResetRef.current = false
-    if (wheelZoomCommitTimeoutRef.current !== null) {
-      window.clearTimeout(wheelZoomCommitTimeoutRef.current)
-      wheelZoomCommitTimeoutRef.current = null
-    }
-
-    if (!liveLayerTransformActiveRef.current) {
-      clearViewportLayerTransform()
-      liveLayerBaseViewportRef.current = renderedViewportRef.current
-      liveLayerTransformActiveRef.current = true
-    }
-
-    event.currentTarget.setPointerCapture(event.pointerId)
-
-    if (event.pointerType === 'touch') {
-      touchPointersRef.current.set(event.pointerId, {
-        pointerId: event.pointerId,
-        clientX: event.clientX,
-        clientY: event.clientY
-      })
-
-      if (touchPointersRef.current.size >= 2) {
-        beginPinchZoom()
-        setOpenNodeId(undefined)
-        setLocalHoveredNodeId(undefined)
-        return
-      }
-    }
-
-    dragRef.current = {
-      active: true,
-      pointerId: event.pointerId,
-      originX: event.clientX,
-      originY: event.clientY,
-      panX: viewportRef.current.panX,
-      panY: viewportRef.current.panY
-    }
-    setIsPanning(true)
-    setOpenNodeId(undefined)
-    setLocalHoveredNodeId(undefined)
-  }
-
-  const handleCanvasPointerMove = (event: ReactPointerEvent<HTMLDivElement>) => {
-    if (event.pointerType === 'touch' && touchPointersRef.current.has(event.pointerId)) {
-      touchPointersRef.current.set(event.pointerId, {
-        pointerId: event.pointerId,
-        clientX: event.clientX,
-        clientY: event.clientY
-      })
-
-      const [first, second] = touchPoints()
-
-      if (first && second) {
-        event.preventDefault()
-
-        if (!pinchZoomRef.current.active) {
-          beginPinchZoom()
-        }
-
-        const currentDistance = Math.max(1, distanceBetweenTouchPoints(first, second))
-        const anchor = midpointRelativeToCanvas(event.currentTarget, first, second)
-        const nextZoom = pinchZoomRef.current.initialZoom * (currentDistance / pinchZoomRef.current.initialDistance)
-
-        zoomAroundPoint(nextZoom, anchor.x, anchor.y, { live: true })
-        return
-      }
-    }
-
-    const drag = dragRef.current
-
-    if (!drag.active || drag.pointerId !== event.pointerId) {
-      return
-    }
-
-    event.preventDefault()
-    const canvasSize = canvasSizeRef.current
-    const minZoom = calculateMaxZoomOut(currentRenderNodes, canvasSize)
-    const nextViewport = clampViewportToPanBounds(
-      currentRenderNodes,
-      canvasSize,
-      {
-        zoom: viewportRef.current.zoom,
-        panX: drag.panX + event.clientX - drag.originX,
-        panY: drag.panY + event.clientY - drag.originY
-      },
-      zoomFocusRef.current,
-      minZoom
-    )
-
-    viewportRef.current = nextViewport
-    scheduleViewportLayerTransform(nextViewport)
-  }
-
-  const stopPanning = (event: ReactPointerEvent<HTMLDivElement>) => {
-    const shouldCommitViewport = dragRef.current.active && dragRef.current.pointerId === event.pointerId
-
-    if (dragRef.current.pointerId === event.pointerId && event.currentTarget.hasPointerCapture(event.pointerId)) {
-      event.currentTarget.releasePointerCapture(event.pointerId)
-    }
-
-    if (event.pointerType === 'touch') {
-      touchPointersRef.current.delete(event.pointerId)
-
-      if (event.currentTarget.hasPointerCapture(event.pointerId)) {
-        event.currentTarget.releasePointerCapture(event.pointerId)
-      }
-    }
-
-    const shouldCommitPinchZoom = pinchZoomRef.current.active && touchPointersRef.current.size < 2
-
-    dragRef.current.active = false
-    dragRef.current.pointerId = null
-
-    if (shouldCommitViewport || shouldCommitPinchZoom) {
-      if (wheelZoomCommitTimeoutRef.current !== null) {
-        window.clearTimeout(wheelZoomCommitTimeoutRef.current)
-        wheelZoomCommitTimeoutRef.current = null
-      }
-
-      pinchZoomRef.current.active = false
-      pendingPanTransformResetRef.current = true
-      setViewport(viewportRef.current, { userControlled: true })
-    }
-
-    if (!shouldCommitPinchZoom && touchPointersRef.current.size < 2) {
-      pinchZoomRef.current.active = false
-    }
-
-    setIsPanning(false)
-  }
-
-  const handleCanvasWheel = useCallback(
-    (event: WheelEvent) => {
-      if (event.deltaY === 0) {
-        return
-      }
-
-      if (event.cancelable) {
-        event.preventDefault()
-      }
-      event.stopPropagation()
-
-      const canvasElement = canvasRef.current
-      if (!canvasElement) {
-        return
-      }
-
-      const rect = canvasElement.getBoundingClientRect()
-      const anchorX = event.clientX - rect.left
-      const anchorY = event.clientY - rect.top
-      const factor = event.deltaY > 0 ? WHEEL_ZOOM_OUT : WHEEL_ZOOM_IN
-
-      zoomAroundPoint(viewportRef.current.zoom * factor, anchorX, anchorY, { live: true })
-    },
-    [zoomAroundPoint]
-  )
-
-  useEffect(() => {
-    const canvasElement = canvasRef.current
-    if (!canvasElement) {
-      return undefined
-    }
-
-    canvasElement.addEventListener('wheel', handleCanvasWheel, { passive: false })
-
-    return () => canvasElement.removeEventListener('wheel', handleCanvasWheel)
-  }, [handleCanvasWheel])
-
-  const handleFullscreen = useCallback(() => {
-    if (onFullscreen) {
-      onFullscreen()
-      return
-    }
-
-    const canvasElement = canvasRef.current
-
-    if (!canvasElement || typeof canvasElement.requestFullscreen !== 'function') {
-      return
-    }
-
-    void canvasElement.requestFullscreen().catch((error: unknown) => {
-      console.warn('Unable to enter mesh fullscreen mode', error)
-    })
-  }, [onFullscreen])
+  const { handleCanvasPointerDown, handleCanvasPointerMove, handleFullscreen, stopPanning } = useMeshVizInteractions({
+    canvasRef,
+    canvasSizeRef,
+    currentRenderNodes,
+    viewportRef,
+    viewportAnimationRef,
+    zoomFocusRef,
+    renderedViewportRef,
+    liveLayerBaseViewportRef,
+    liveLayerTransformActiveRef,
+    wheelZoomCommitTimeoutRef,
+    pendingPanTransformResetRef,
+    hasUserControlledViewportRef,
+    dragRef,
+    touchPointersRef,
+    pinchZoomRef,
+    setIsPanning,
+    setOpenNodeId,
+    setLocalHoveredNodeId,
+    setViewport,
+    clearViewportLayerTransform,
+    scheduleViewportLayerTransform,
+    zoomAroundPoint,
+    onFullscreen
+  })
 
   const screenLinks = links.map((link) => ({
     ...link,
@@ -1811,309 +898,78 @@ export const MeshViz = forwardRef<MeshVizHandle, MeshVizProps>(function MeshViz(
   const viewportControlIconClassName = isFullscreen ? 'size-6' : 'size-3'
 
   return (
-    <section className="panel-shell flex h-full min-h-0 flex-col overflow-hidden rounded-[var(--radius-lg)] border border-border bg-panel">
-      <header className="flex shrink-0 items-center justify-between border-b border-border-soft px-4 py-3">
-        <h2 className="type-panel-title">Mesh overview</h2>
-        {!compact && (
-          <button
-            onClick={handleFullscreen}
-            type="button"
-            className="ui-control inline-flex items-center gap-1.5 rounded-[var(--radius)] border px-2.5 py-1 text-[length:var(--density-type-caption)] font-medium"
-          >
-            <Maximize2 className="size-3" /> Fullscreen
-          </button>
-        )}
-      </header>
-      <div className="flex min-h-0 flex-1 p-3.5">
-        <div
-          ref={canvasRef}
-          data-testid="mesh-canvas"
-          className={cn(
-            'relative w-full touch-none overflow-hidden rounded-[var(--radius-lg)] mesh-canvas',
-            isPanning ? 'cursor-grabbing' : 'cursor-grab'
-          )}
-          style={{
-            height: height ?? '100%',
-            background:
-              'radial-gradient(ellipse at 60% 40%, color-mix(in oklab, var(--color-accent) 10%, var(--color-panel-strong)) 0%, var(--color-panel-strong) 60%, var(--color-panel) 100%)'
-          }}
-          onPointerDown={handleCanvasPointerDown}
-          onPointerMove={handleCanvasPointerMove}
-          onPointerUp={stopPanning}
-          onPointerCancel={stopPanning}
-        >
-          <div className="absolute left-3.5 top-3 z-10 flex flex-wrap items-center gap-2.5 font-mono text-[length:var(--density-type-label)] uppercase tracking-[0.14em] text-muted-foreground">
-            <span className="inline-flex items-center gap-1.5 rounded-full border border-border bg-panel/90 px-2 py-px text-accent">
-              <span className="size-[5px] rounded-full bg-current mesh-live-pulse" /> Live
-            </span>
-            <span>
-              {nodes.length} nodes{debugNodes.length > 0 ? ` + ${debugNodes.length} debug` : ''} · {linkCount} links ·
-              Nearest mesh
-            </span>
-          </div>
-
-          {isDevelopment && (
-            <div
-              data-testid="mesh-max-zoom-label"
-              className="pointer-events-none absolute right-3.5 top-3 z-10 rounded-full border border-border bg-panel/90 px-2 py-px font-mono text-[length:var(--density-type-label)] uppercase tracking-[0.14em] text-muted-foreground"
-            >
-              Max Zoom: {maxZoomOutLabel}
-            </div>
-          )}
-
-          <svg
-            viewBox={`0 0 ${safeCanvasWidth} ${safeCanvasHeight}`}
-            preserveAspectRatio="none"
-            className="pointer-events-none absolute inset-0 h-full w-full overflow-hidden"
-            role="img"
-            aria-label="Nearest mesh topology"
-          >
-            <defs>
-              <pattern
-                ref={gridPatternRef}
-                id="mesh-viz-grid"
-                width={gridSize}
-                height={gridSize}
-                patternUnits="userSpaceOnUse"
-                patternTransform={gridTransform}
-              >
-                {gridMode === 'line' ? (
-                  <path
-                    ref={gridPathRef}
-                    data-testid="mesh-viz-line-grid"
-                    d={`M ${gridSize} 0 L 0 0 0 ${gridSize}`}
-                    fill="none"
-                    stroke="color-mix(in oklab, var(--color-foreground) 7.2%, transparent)"
-                    strokeWidth="1"
-                  />
-                ) : (
-                  <>
-                    <circle
-                      ref={gridDotRef}
-                      data-testid="mesh-viz-dot-grid"
-                      cx="0"
-                      cy="0"
-                      r="1.35"
-                      fill={dotColorScheme.colors[0]}
-                    />
-                    <circle
-                      ref={gridAccentDotRef}
-                      data-testid="mesh-viz-accent-dot-grid"
-                      cx={gridSize / 2}
-                      cy={gridSize / 2}
-                      r="1.25"
-                      fill={dotColorScheme.colors[1]}
-                    />
-                    <circle
-                      ref={gridTertiaryDotRef}
-                      data-testid="mesh-viz-tertiary-dot-grid"
-                      cx="0"
-                      cy={gridSize / 2}
-                      r="0.85"
-                      fill={dotColorScheme.colors[2]}
-                    />
-                  </>
-                )}
-              </pattern>
-            </defs>
-            <rect width={safeCanvasWidth} height={safeCanvasHeight} fill="url(#mesh-viz-grid)" />
-            <g ref={svgPanLayerRef}>
-              {screenLinks.map((link) => (
-                <line
-                  key={link.id}
-                  className="mesh-link"
-                  data-link-lifecycle={linkLifecyclePhase(link.source.id, link.target.id)}
-                  data-mesh-link-id={link.id}
-                  data-source-node-id={link.source.id}
-                  data-target-node-id={link.target.id}
-                  data-testid="mesh-link"
-                  pathLength={1}
-                  x1={link.sourcePoint.x}
-                  y1={link.sourcePoint.y}
-                  x2={link.targetPoint.x}
-                  y2={link.targetPoint.y}
-                  stroke="color-mix(in oklab, var(--color-accent) 48%, var(--color-border))"
-                  strokeDasharray="0.0275 0.0275"
-                  strokeLinecap="round"
-                  strokeWidth="1"
-                  opacity={link.dimmed ? '0.18' : '0.62'}
-                  vectorEffect="non-scaling-stroke"
-                />
-              ))}
-              {isDevelopment && showPanBounds && nodeBoundsRect && deadZoneRect && centeredBoundsRect && (
-                <g aria-label="Mesh pan bounds debug overlay">
-                  <rect
-                    data-testid="mesh-pan-dead-zone-box"
-                    x={deadZoneRect.x}
-                    y={deadZoneRect.y}
-                    width={deadZoneRect.width}
-                    height={deadZoneRect.height}
-                    fill="color-mix(in oklab, var(--color-accent) 7%, transparent)"
-                    stroke="color-mix(in oklab, var(--color-accent) 72%, transparent)"
-                    strokeDasharray="8 6"
-                    strokeWidth="1.2"
-                    vectorEffect="non-scaling-stroke"
-                  />
-                  <rect
-                    data-testid="mesh-node-bounds-box"
-                    x={nodeBoundsRect.x}
-                    y={nodeBoundsRect.y}
-                    width={nodeBoundsRect.width}
-                    height={nodeBoundsRect.height}
-                    fill="none"
-                    stroke="color-mix(in oklab, var(--color-good) 78%, transparent)"
-                    strokeWidth="1.4"
-                    vectorEffect="non-scaling-stroke"
-                  />
-                  <rect
-                    data-testid="mesh-centered-bounds-box"
-                    x={centeredBoundsRect.x}
-                    y={centeredBoundsRect.y}
-                    width={centeredBoundsRect.width}
-                    height={centeredBoundsRect.height}
-                    fill="none"
-                    stroke="color-mix(in oklab, var(--color-warn) 82%, transparent)"
-                    strokeDasharray="6 5"
-                    strokeWidth="1.2"
-                    vectorEffect="non-scaling-stroke"
-                  />
-                </g>
-              )}
-            </g>
-          </svg>
-
-          <div
-            ref={packetLayerRef}
-            className="pointer-events-none absolute inset-0 z-[5]"
-            data-testid="mesh-packet-layer"
-            style={{ transformOrigin: '0 0', willChange: 'transform' }}
-            aria-hidden="true"
-          />
-
-          <div
-            ref={nodeLayerRef}
-            className="absolute inset-0"
-            style={{ transformOrigin: '0 0', willChange: 'transform' }}
-          >
-            {renderNodes.map((node) => {
-              const peer = getNodePeer?.(node)
-              const isDimmed = dimmedNodeIds?.has(node.id) ?? false
-
-              return (
-                <div key={node.id} style={{ opacity: isDimmed ? 0.24 : 1, transition: 'opacity 180ms ease' }}>
-                  <MeshVizNode
-                    node={node}
-                    peer={peer}
-                    selfId={selfId}
-                    selectedNodeId={selectedNodeId}
-                    openNodeId={openNodeId}
-                    hoveredNodeId={hoveredNodeId}
-                    shouldFadeNodeLabels={shouldFadeNodeLabels}
-                    reduceMotion={reduceMotion}
-                    canvasWidth={safeCanvasWidth}
-                    canvasHeight={safeCanvasHeight}
-                    viewport={viewport}
-                    nodeColors={dotColorScheme.nodeColors}
-                    lifecycle={nodeLifecyclePhase(node.id)}
-                    radarPingRef={radarPingRef}
-                    onHoverStart={setLocalHoveredNodeId}
-                    onHoverEnd={(nodeId) =>
-                      setLocalHoveredNodeId((current) => (current === nodeId ? undefined : current))
-                    }
-                    onToggleOpen={(nodeId) => {
-                      setOpenNodeId((current) => (current === nodeId ? undefined : nodeId))
-                    }}
-                    onCloseOpen={() => setOpenNodeId(undefined)}
-                  />
-                </div>
-              )
-            })}
-          </div>
-
-          <div
-            ref={labelLayerRef}
-            className="pointer-events-none absolute inset-0 z-[40]"
-            data-testid="mesh-node-label-layer"
-            style={{ transformOrigin: '0 0', willChange: 'transform' }}
-            aria-hidden="true"
-          >
-            {renderNodes.map((node) => {
-              const peer = getNodePeer?.(node)
-              const isDimmed = dimmedNodeIds?.has(node.id) ?? false
-
-              return (
-                <div key={node.id} style={{ opacity: isDimmed ? 0.24 : 1, transition: 'opacity 180ms ease' }}>
-                  <MeshVizNodeLabel
-                    node={node}
-                    peer={peer}
-                    selfId={selfId}
-                    selectedNodeId={selectedNodeId}
-                    openNodeId={openNodeId}
-                    hoveredNodeId={hoveredNodeId}
-                    shouldFadeNodeLabels={shouldFadeNodeLabels}
-                    reduceMotion={reduceMotion}
-                    canvasWidth={safeCanvasWidth}
-                    canvasHeight={safeCanvasHeight}
-                    viewport={viewport}
-                    nodeColors={dotColorScheme.nodeColors}
-                    lifecycle={nodeLifecyclePhase(node.id)}
-                  />
-                </div>
-              )
-            })}
-          </div>
-
-          {isDevelopment && (
-            <MeshVizDebugControls
-              debugNodeCount={debugNodes.length}
-              dotColorSchemeIndex={dotColorSchemeIndex}
-              dotColorSchemes={dotColorSchemes}
-              gridMode={gridMode}
-              isFullscreen={isFullscreen}
-              onAddDebugNode={addDebugNode}
-              onDotColorSchemeChange={selectDotColorScheme}
-              onDotColorSchemeNext={cycleDotColorScheme}
-              onGridModeChange={setGridMode}
-              onPlayRandomTraffic={playRandomTraffic}
-              onPlaySelfTraffic={playSelfTraffic}
-              onRemoveDebugNode={removeDebugNode}
-              onShowPanBoundsChange={setShowPanBounds}
-              showPanBounds={showPanBounds}
-            />
-          )}
-
-          <div className="absolute bottom-3 right-3 flex flex-col gap-1.5">
-            <button
-              type="button"
-              className={viewportControlClassName}
-              aria-label="Zoom in"
-              onPointerDown={(event) => event.stopPropagation()}
-              onClick={() => zoomAtCenter(BUTTON_ZOOM_IN)}
-            >
-              <Plus className={viewportControlIconClassName} />
-            </button>
-            <button
-              type="button"
-              className={viewportControlClassName}
-              aria-label="Zoom out"
-              onPointerDown={(event) => event.stopPropagation()}
-              onClick={() => zoomAtCenter(BUTTON_ZOOM_OUT)}
-            >
-              <Minus className={viewportControlIconClassName} />
-            </button>
-            <button
-              type="button"
-              className={viewportControlClassName}
-              aria-label="Reset view"
-              onPointerDown={(event) => event.stopPropagation()}
-              onClick={fitNodes}
-            >
-              <RotateCcw className={viewportControlIconClassName} />
-            </button>
-          </div>
-        </div>
-      </div>
-    </section>
+    <MeshVizCanvas
+      canvasRef={canvasRef}
+      gridPatternRef={gridPatternRef}
+      gridPathRef={gridPathRef}
+      gridDotRef={gridDotRef}
+      gridAccentDotRef={gridAccentDotRef}
+      gridTertiaryDotRef={gridTertiaryDotRef}
+      svgPanLayerRef={svgPanLayerRef}
+      nodeLayerRef={nodeLayerRef}
+      labelLayerRef={labelLayerRef}
+      packetLayerRef={packetLayerRef}
+      radarPingRef={radarPingRef}
+      safeCanvasWidth={safeCanvasWidth}
+      safeCanvasHeight={safeCanvasHeight}
+      gridSize={gridSize}
+      gridTransform={gridTransform}
+      gridMode={gridMode}
+      dotColorScheme={dotColorScheme}
+      screenLinks={screenLinks}
+      isDevelopment={isDevelopment}
+      showPanBounds={showPanBounds}
+      nodeBoundsRect={nodeBoundsRect}
+      deadZoneRect={deadZoneRect}
+      centeredBoundsRect={centeredBoundsRect}
+      isPanning={isPanning}
+      isFullscreen={isFullscreen}
+      height={height}
+      compact={compact}
+      nodes={nodes}
+      debugNodes={debugNodes}
+      linkCount={linkCount}
+      maxZoomOutLabel={maxZoomOutLabel}
+      renderNodes={renderNodes}
+      dimmedNodeIds={dimmedNodeIds}
+      selfId={selfId}
+      selectedNodeId={selectedNodeId}
+      openNodeId={openNodeId}
+      hoveredNodeId={hoveredNodeId}
+      shouldFadeNodeLabels={shouldFadeNodeLabels}
+      reduceMotion={reduceMotion}
+      viewport={viewport}
+      nodeLifecyclePhase={nodeLifecyclePhase}
+      linkLifecyclePhase={linkLifecyclePhase}
+      getNodePeer={getNodePeer}
+      dotColorSchemeIndex={dotColorSchemeIndex}
+      dotColorSchemes={dotColorSchemes}
+      onPointerDown={handleCanvasPointerDown}
+      onPointerMove={handleCanvasPointerMove}
+      onPointerUp={stopPanning}
+      onPointerCancel={stopPanning}
+      onFullscreen={handleFullscreen}
+      onAddDebugNode={addDebugNode}
+      onRemoveDebugNode={removeDebugNode}
+      onDotColorSchemeChange={selectDotColorScheme}
+      onDotColorSchemeNext={cycleDotColorScheme}
+      onGridModeChange={setGridMode}
+      onPlayRandomTraffic={playRandomTraffic}
+      onPlaySelfTraffic={playSelfTraffic}
+      onShowPanBoundsChange={setShowPanBounds}
+      onZoomAtCenter={zoomAtCenter}
+      onFitNodes={fitNodes}
+      onNodeHoverStart={setLocalHoveredNodeId}
+      onNodeHoverEnd={(nodeId) => {
+        setLocalHoveredNodeId((current) => (current === nodeId ? undefined : current))
+      }}
+      onNodeToggleOpen={(nodeId) => {
+        setOpenNodeId((current) => (current === nodeId ? undefined : nodeId))
+      }}
+      onNodeCloseOpen={() => setOpenNodeId(undefined)}
+      viewportControlClassName={viewportControlClassName}
+      viewportControlIconClassName={viewportControlIconClassName}
+    />
   )
 })
